+++ commands/proposal_submission.go
@@ -79,10 +79,11 @@ func checkProposalSubmission(cmd *commandspb.ProposalSubmission) Errors {
 		)
 	}
 
+	// check for enactment timestamp
 	switch cmd.Terms.Change.(type) {
 	case *types.ProposalTerms_NewFreeform:
 		if cmd.Terms.EnactmentTimestamp != 0 {
-			errs.AddForProperty("proposal_submission.terms.enactment_timestamp", ErrIsNotValid)
+			errs.AddForProperty("proposal_submission.terms.enactment_timestamp", ErrIsNotSupported)
 		}
 	default:
 		if cmd.Terms.EnactmentTimestamp <= 0 {
@@ -93,6 +94,23 @@ func checkProposalSubmission(cmd *commandspb.ProposalSubmission) Errors {
 			errs.AddForProperty("proposal_submission.terms.closing_timestamp",
 				errors.New("cannot be after enactment time"),
 			)
+		}
+	}
+
+	// check for validation timestamp
+	switch cmd.Terms.Change.(type) {
+	case *types.ProposalTerms_NewAsset:
+		if cmd.Terms.ValidationTimestamp == 0 {
+			errs.AddForProperty("proposal_submission.terms.validation_timestamp", ErrMustBePositive)
+		}
+		if cmd.Terms.ValidationTimestamp > cmd.Terms.ClosingTimestamp {
+			errs.AddForProperty("proposal_submission.terms.validation_timestamp",
+				errors.New("cannot be after closing time"),
+			)
+		}
+	default:
+		if cmd.Terms.ValidationTimestamp != 0 {
+			errs.AddForProperty("proposal_submission.terms.validation_timestamp", ErrIsNotSupported)
 		}
 	}